docs(icons): cleanup stories
