+++ packages/core/src/components/Icon/__stories__/Icon.mdx
@@ -1,5 +1,4 @@
 import { Meta } from "@storybook/blocks";
-import "./Icon.stories.module.scss";
 import * as IconStories from "./Icon.stories";
 
 <Meta of={IconStories} />
