+++ packages/core/src/components/Icon/__stories__/Icon.stories.tsx
@@ -1,10 +1,13 @@
-import React from "react";
+import React, { useState } from "react";
 import Icon from "../Icon";
 import { Bolt } from "@vibe/icons";
 import { createStoryMetaSettingsDecorator } from "../../../storybook";
 import { createComponentTemplate } from "vibe-storybook-components";
-import IconsList from "./IconsList";
-import "./Icon.stories.module.scss";
+import Search from "../../Search/Search";
+import * as allIcons from "@vibe/icons";
+import Flex from "../../Flex/Flex";
+import iconsMetaData from "@vibe/icons/meta";
+import { SubIcon } from "src/types";
 
 const metaSettings = createStoryMetaSettingsDecorator({
   component: Icon,
@@ -57,7 +60,6 @@ export const CustomSvg = {
       iconType="src"
       icon="https://cdn.monday.com/images/apps/custom-icons/Form.svg"
       iconLabel="my src awesome icon"
-      className="icon-story-custom-icon"
       useCurrentColor
     />
   )
@@ -84,7 +86,42 @@ export const Color = {
 };
 
 export const IconsListStory = {
-  render: IconsList.bind({}),
-  name: "Icons List",
-  args: {}
+  render: () => {
+    interface IconMeta {
+      name: string;
+      tags: string;
+      file: string;
+    }
+    const [query, setQuery] = useState("");
+    return (
+      <section style={{ width: "100%" }}>
+        <Search value={query} onChange={setQuery} placeholder="Search for icons" />
+        <div
+          style={{
+            display: "grid",
+            gridTemplateColumns: "repeat(auto-fill, minmax(250px, 1fr))",
+            gap: "var(--sb-spacing-large)",
+            marginTop: "var(--sb-spacing-large)"
+          }}
+        >
+          {iconsMetaData
+            .filter((icon: IconMeta) => {
+              return `${icon.tags},${icon.name}`.toLowerCase().includes(query.toLowerCase());
+            })
+            .map((icon: IconMeta) => {
+              const fileName = icon.file.split(".")[0] as string;
+              const Component = allIcons[fileName as keyof typeof allIcons] as SubIcon;
+              return (
+                <>
+                  <Flex style={{ color: "var(--sb-icon-color)" }} gap="small">
+                    <Icon icon={Component} iconSize={26} />
+                    <span>{icon.name}</span>
+                  </Flex>
+                </>
+              );
+            })}
+        </div>
+      </section>
+    );
+  }
 };
